New, more generic AssetCache class with AssetInfo struct
and new MusicManager class to tag along. Both of these are not global (AssetCache can be. There is a global one but it will probably be seldom used). MusicManager is generally limited to there being one, in the current level. This is not a hard rule as it is not a singleton, but you probably don't want more than one piece of music playing at the same time. Multiple could be used to fade between music though. Maybe a MusicFader class or something