Fixed level loading and entity destruction bugs
How you ask? I did something I shoulda done a long time ago... I just made them happen at the end of the current frame. So if any entity is pending destroy (set to true from destroy() method), then it is erased from the entities vector in the level. and if a new level is set to open (created from createAndOpenLevel() method), if it's valid, then it opens it at the end of the frame, thus not having anything to do with input or updates. It is done separately!