Removed chaiscript, favouring class registry instead
Very simple reflection with serialising to and from json,
added Vector 2 struct to wrap and be a helper for SFML vector 2.
CURRENTLY COMPILE ERROR. Not really a clue as to why but working on it!!!
Also I like D Lang a lot, want to try doing more stuff in that.