Changed loadResource catch to std::exception
It could be a bad_alloc too, which is neither an sf::Exception nor an std::runtime_error.
And, since sf::Exception does nothing more than be its own version of std::runtime_error, I see no reason to use it anyway. Unless more is added to it, and I need to use those things, std::exception is fine.