Refactored components
An entity now has a root component.
By default, the root component is the base CComponent.
An entity's root component MUST be valid.
Components can have children, and a parent.
Each component owns their children.
A component's outer is still the entity that owns it.
Components are added to an entity's root component via the "getRootComponent" function (this will probably change to general "addComponent", etc. functions in the entity which add components to the root component)