Rejigged entity components
There's the base CComponent, which only ticks, it does not draw nor have a transform. CTransformComponent is the base of all drawable, transformable, components like sprites, etc.
These are heavily inspired by Unreal Engine's UActorComponent and UPrimitiveComponent respectively.
An entity has a vector of its components, and that entity owns and manages the lifetime of its components. Transformable components can have parents and children, both are non-owning. In order to use these effectively, you first create the component in the entity (in its constructor typically), then set its parent if it is a transformable component and is not the root. You should set the root component also, as at the moment it does not make one for you