Textures now point to SDL_Texture internally
This is so SDL_Textures don't need to be created and destroyed every frame (slow!), they're created once for each allocated CTexture object, and destroyed when that CTexture object is deleted from its shared_ptr.
Also the naming of getPixelCount I'm not sure on. It's not how many pixels there are, it's that * channels. getSize is in used, might change that to texture size, and then have a getPixelSize, idk